third_party/blink/web_tests/platform/win11-arm64/http/tests/security/script-crossorigin-loads-correctly-credentials-expected.txt

CONSOLE ERROR: Access to script at 'http://localhost:8000/security/resources/cors-script.php?credentials=true' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
This test passes if the script loads correctly.

FAIL
